requirements.txt
fastapi
uvicorn
google-generativeai
fpdf2
pydantic